+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Fix content range format in multipart strategy
 - Fix hash calculation (tree hash) of the uploaded archive
+- Check if the audit file exists when listing it
 
 ### Added
 - Verifies the hash of each uploaded part in multipart strategy